CHC solver: extract inductiveness check
The inductiveness check is complex, and now extracted into a separate .cpp
file.
